Added basic animations
Added a stretch and jump animation for the adoption scene. Added an idle sitting and walking animation to the cat for the living room scene. NOTE: The idle sitting animation has a weird tail issue that will be resolved soon. These are super basic and set in place for testing purposes and will be tweaked to look amazing later!